Code Review Fix for DMS TripController
