+++ DMS.API/Web.config
@@ -52,8 +52,8 @@
     <add key="TMSLogin" value="admin" />
     <add key="TMSPassword" value="admin" />
     <add key="AllowPushNotifications" value="true" />
-    <add key="FCM_Authorization" value="key=AAAASlr9K0Y:APA91bEpVYc3duC6fq7w38AGljxFkKXMMUHrSMF18fXRBHnVU-RWWyymfrGRX3D0H_zwflmJtjAtrcBosSt4sQxvrkuUZ6wjWP2ZvGwWzyxnYYjue-sIs7466OA1DWL6_X4UJ5gZcgPD"/>
-    <add key="ShipmentListOCRURL" value="http://18.138.206.192:4003/extract"/>
+    <add key="FCM_Authorization" value="key=AAAASlr9K0Y:APA91bEpVYc3duC6fq7w38AGljxFkKXMMUHrSMF18fXRBHnVU-RWWyymfrGRX3D0H_zwflmJtjAtrcBosSt4sQxvrkuUZ6wjWP2ZvGwWzyxnYYjue-sIs7466OA1DWL6_X4UJ5gZcgPD" />
+    <add key="ShipmentListOCRURL" value="http://18.138.206.192:4003/extract" />
   </appSettings>
   <system.web>
     <compilation debug="true" targetFramework="4.6.1" />
